💄 [Styling] image 파일 경로 수정 및 Header 스타일링 수정
