+++ src/app/(main)/ProductIntro.tsx
@@ -1,6 +1,5 @@
 import Image from 'next/image';
-import mobileImg from '@/assets/mobileMockup.svg';
-import { Button } from '../ui/button';
+import { Button } from '../../components/ui/button';
 
 export default function ProductIntro() {
     return (
@@ -9,7 +8,13 @@ export default function ProductIntro() {
                 <p className='text-3xl font-bold'>우리 아이를 위한</p>
                 <p className='text-2xl'>맞춤형 성장 패키지</p>
             </div>
-            <Image className='mx-auto' src={mobileImg} alt='' />
+            <Image
+                className='mx-auto'
+                src={'/mobileMockup.svg'}
+                width={190}
+                height={384}
+                alt=''
+            />
             <Button
                 variant={'default'}
                 size={'lg'}
