+++ src/app/(products)/list/MonthAvatar.tsx
@@ -1,5 +1,4 @@
 import Image from 'next/image';
-import avartatImg from '@/assets/sample_avatar.svg';
 
 export default function MonthAvatar() {
     return (
@@ -8,7 +7,8 @@ export default function MonthAvatar() {
                 <Image
                     className='absolute top-[5px] left-2.5'
                     width={50}
-                    src={avartatImg}
+                    height={50}
+                    src={'/sample_avatar.svg'}
                     alt=''
                 />
             </div>
