--- src/components/MonthCategorySelector.tsx
@@ -1,3 +1,0 @@
-export default function MonthCategorySelector() {
-  return <div>MonthCategorySelector</div>;
-}
