--- src/components/MonthCategorySelectorItem.tsx
@@ -1,3 +1,0 @@
-export default function MonthCategorySelectorItem() {
-  return <div>MonthCategorySelectorItem</div>;
-}
